@@ -2,11 +2,14 @@ library flutter_slidable;
 
 export 'src/action_pane_motions.dart';
 export 'src/actions.dart';
+export 'src/auto_close_behavior.dart'
+    show SlidableAutoCloseBehavior, SlidableAutoCloseNotification;
 export 'src/controller.dart'
     show ResizeRequest, SlidableController, ActionPaneType;
 export 'src/dismissible_pane.dart';
 export 'src/dismissible_pane_motions.dart';
-export 'src/notifications.dart'
+export 'src/notifications.dart';
+export 'src/notifications_old.dart'
     show
         SlidableNotification,
         SlidableRatioNotification,
